refactor(vizij-authoring): move pose group inspector into inspector sidebar
