+++ apps/vizij-authoring/src/types/poseGroupInspector.ts
@@ -1,0 +1,6 @@
+export interface PoseGroupInspectorSelection {
+  groupPath: string;
+  label: string;
+  poseIds: string[];
+  nodeId: string;
+}
